Move economy code to chat-plugins/economy.js
