+++ .eslintrc
@@ -5,7 +5,8 @@
 	"globals": {
 		"Config": false, "Monitor": false, "toId": false, "Tools": false, "LoginServer": false,
 		"Users": false, "Rooms": false, "Verifier": false, "CommandParser": false, "Simulator": false,
-		"Tournaments": false, "Dnsbl": false, "Cidr": false, "Sockets": false, "TeamValidator": false, "Ladders": false
+		"Tournaments": false, "Dnsbl": false, "Cidr": false, "Sockets": false, "TeamValidator": false, "Ladders": false,
+		"Economy": false, "Gold": false
 	},
 	"rules": {
 		"comma-dangle": [2, "always-multiline"],
