+++ chat-plugins/roomshop.js
@@ -80,7 +80,7 @@ exports.commands = {
 				if (!RS[item]) return this.errorReply("This item is not in the room shop. Check spelling?");
 				item = RS[item][0];
 				var price = RS[item][2];
-				if (economy.readMoney(user.userid) < price) return this.errorReply("You do not have enough bucks to buy " + item + ". You need " + (price - economy.readMoney(user.userid)) + " more bucks to buy this item.");
+				if (Economy.readMoneySync(user.userid) < price) return this.errorReply("You do not have enough bucks to buy " + item + ". You need " + (price - Economy.readMoneySync(user.userid)) + " more bucks to buy this item.");
 				this.parse('/tb ' + room.founder + ', ' + price);
 				room.add("|raw|<b><u>Room Shop</u>: " + getName(user.name) + "</b> has bought a(n) <u>" + Tools.escapeHTML(item) + "</u> from the room shop for " + price + " buck" + (price > 1 ? "s" : "") + ".").update();
 				this.privateModCommand("(" + user.name + " has bought a(n) " + item + " from the room shop.)");
