+++ tournaments/index.js
@@ -745,7 +745,7 @@ class Tournament {
 
 			try {
 				this.room.add('|raw|<b><font color=' + Gold.hashColor(toId(winner)) + '>' + Tools.escapeHTML(winner) + '</font> has won <font color=#24678d>' + money + '</font> bucks for winning the tournament!</b>');
-				economy.writeMoney('money', toId(winner), money);
+				Economy.writeMoney(toId(winner), money);
 			} catch (e) {}
 		}
 		if (this.autoDisqualifyTimer) clearTimeout(this.autoDisqualifyTimer);
